This is the Customer APP for TML.
[09/08/2015] Added a couresel which doesnt have any picture yet
Added a picture in view/show which is a banana but is itended to be the picture for each customer but will fix that in a later version.
Added A new Model For Tracking with tracking_number and logistic as fields
Added in controller of tracking that it belongs_to :customer
and in customer  has_many :trackings

[09/07/2015]
List is generated in Ascending order based on first name.

What Works:
1. Creating Customers
2. Viewing Customers
3. Destroying Customers
4. Search for Customers based on First Name

What Does Not Work:
1. The Nav Bar that is Fixed on top
2. Nor any of the links/buttons on the top